% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_radial_data.R
\name{analyze_radial_data}
\alias{analyze_radial_data}
\title{Analyze output from my CellProfiler micropattern project}
\usage{
analyze_radial_data(input_file, DNA_channel = "DNA")
}
\arguments{
\item{input_file}{The csv output file from my micropattern project.}

\item{DNA_channel}{Specify the name of the DNA channel if different from DNA.}
}
\description{
Analyze output from my CellProfiler micropattern project
}
