% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ideo_data.R
\name{geom_ideogram}
\alias{geom_ideogram}
\title{Pull and organise chromosome ideogram data from HGSC}
\usage{
geom_ideogram(
  chromosome,
  genome = "hg38",
  Y_offset = 0,
  height = 1,
  label = FALSE
)
}
\arguments{
\item{chromosome}{Chromosome to display by number. Typically autosomes are listed first in numerical order, then non-homologous chromosomes.}

\item{genome}{Abbreviated genome name for data to be pulled}

\item{Y_offset}{y-coordinate for the centre of the ideogram}

\item{height}{Height of the ideogram}
}
\description{
Pull and organise chromosome ideogram data from HGSC
}
