% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_theme.R
\name{lb_theme}
\alias{lb_theme}
\title{Basic theme for ggplots}
\usage{
lb_theme(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  dark_mode = FALSE
)
}
\arguments{
\item{base_size}{Size that various elements use.}

\item{base_family}{Font family to use.}

\item{base_line_size}{Defaults to base_size / 22.}

\item{base_rect_size}{Defaults to base_size / 22.}

\item{dark_mode}{Switch to dark mode (black background, white lines and text).}
}
\description{
Basic theme for ggplots
}
